chore: refactor tx fuzzer
