+++ barretenberg/cpp/src/barretenberg/avm_fuzzer/fuzz_lib/fuzz.cpp
@@ -140,8 +140,8 @@ SimulatorResult fuzz(FuzzerData& fuzzer_data)
         fuzz_info("JsSimulator  ", js_result);
         throw std::runtime_error("Simulator results are different");
     }
-    // fuzz_info("Simulator results match successfully");
-    // fuzz_info("CppSimulator ", cpp_result);
-    // fuzz_info("JsSimulator  ", js_result);
+    fuzz_info("Simulator results match successfully");
+    fuzz_info("CppSimulator ", cpp_result);
+    fuzz_info("JsSimulator  ", js_result);
     return cpp_result;
 }
